refactor: search페이지 api호출 폴더 이동
